:bug: Fix : 웹소켓 https proxy 설정
프론트 배포 서버에서 websocket 연결 오류 수정
